API register

+++ app/Http/Controllers/ProdiController.php
@@ -39,6 +39,9 @@ class prodiController extends Controller{
         //dump($request);
         //echo $request->nama;
 
+        $this->authorize('create', Prodi::class);
+
+
         $validateData = $request->validate([
             'nama' => 'required|min:5|max:20',
             'foto' => 'required|file|image|max:5000',
@@ -64,6 +67,8 @@ class prodiController extends Controller{
      }
 
      public function update(Request $request, Prodi $prodi){
+
+        $this->authorize('update', $prodi);
         $validateData = $request->validate([
             'nama'=> 'required|min:5|max:20',
         ]);
@@ -78,7 +83,10 @@ class prodiController extends Controller{
      }
 
      public function destroy(Prodi $prodi){
+        
+        $this->authorize('delete', $prodi);
         $prodi->delete();
+
         return redirect()->route('prodi.index')->with("info". "Prodi $prodi->nama berhasil dihapus.");
      }
